chore: upgraded cdk
added pipeline stack
