+++ cdk.json
@@ -1,3 +1,9 @@
 {
-  "app": "npx ts-node --prefer-ts-exts src/main.ts"
+  "app": "npx ts-node --prefer-ts-exts src/main.ts",
+  "context": {
+    "@aws-cdk/core:newStyleStackSynthesis": "true",
+    "whitelist": [
+      "87.123.55.147/32"
+    ]
+  }
 }
