+++ package.json
@@ -30,7 +30,7 @@
     "@types/node": "^10.17.0",
     "@typescript-eslint/eslint-plugin": "^4.3.0",
     "@typescript-eslint/parser": "^4.3.0",
-    "aws-cdk": "^1.87.1",
+    "aws-cdk": "^1.88.0",
     "eslint": "^7.16.0",
     "eslint-import-resolver-node": "^0.3.4",
     "eslint-import-resolver-typescript": "^2.3.0",
@@ -38,22 +38,26 @@
     "jest": "^26.4.2",
     "jest-junit": "^12",
     "json-schema": "^0.2.5",
-    "projen": "^0.15.14",
+    "projen": "^0.16.2",
     "standard-version": "^9.0.0",
     "ts-jest": "^26.1.0",
     "ts-node": "^9.1.1",
     "typescript": "^3.9.5"
   },
   "dependencies": {
-    "@aws-cdk/assert": "^1.87.1",
-    "@aws-cdk/aws-apigateway": "^1.87.1",
-    "@aws-cdk/aws-iam": "^1.87.1",
-    "@aws-cdk/aws-lambda": "^1.87.1",
-    "@aws-cdk/aws-lambda-event-sources": "^1.87.1",
-    "@aws-cdk/aws-logs": "^1.87.1",
-    "@aws-cdk/aws-s3": "^1.87.1",
-    "@aws-cdk/aws-sqs": "^1.87.1",
-    "@aws-cdk/core": "^1.87.1"
+    "@aws-cdk/assert": "^1.88.0",
+    "@aws-cdk/aws-apigateway": "^1.88.0",
+    "@aws-cdk/aws-codebuild": "^1.88.0",
+    "@aws-cdk/aws-codepipeline": "^1.88.0",
+    "@aws-cdk/aws-codepipeline-actions": "^1.88.0",
+    "@aws-cdk/aws-iam": "^1.88.0",
+    "@aws-cdk/aws-lambda": "^1.88.0",
+    "@aws-cdk/aws-lambda-event-sources": "^1.88.0",
+    "@aws-cdk/aws-logs": "^1.88.0",
+    "@aws-cdk/aws-s3": "^1.88.0",
+    "@aws-cdk/aws-sqs": "^1.88.0",
+    "@aws-cdk/core": "^1.88.0",
+    "@aws-cdk/pipelines": "^1.88.0"
   },
   "bundledDependencies": [],
   "keywords": [
